+++ packages/editor/src/components/collaborators-presence/index.tsx
@@ -1,4 +1,7 @@
-import { Button } from '@wordpress/components';
+import {
+	Button,
+	privateApis as componentsPrivateApis,
+} from '@wordpress/components';
 import { useState } from '@wordpress/element';
 import {
 	privateApis,
@@ -6,12 +9,16 @@ import {
 } from '@wordpress/core-data';
 import { __, sprintf } from '@wordpress/i18n';
 
-import { Avatar } from './avatar';
 import { CollaboratorsList } from './list';
 import { unlock } from '../../lock-unlock';
+import { getAvatarUrl } from '../collaborators-overlay/get-avatar-url';
+import { getAvatarBorderColor } from '../collab-sidebar/utils';
 
 import './styles/collaborators-presence.scss';
+import { CollaboratorsOverlay } from '../collaborators-overlay';
+
 const { useActiveCollaborators } = unlock( privateApis );
+const { Avatar, AvatarGroup } = unlock( componentsPrivateApis );
 
 interface CollaboratorsPresenceProps {
 	postId: number | null;
@@ -22,9 +29,9 @@ interface CollaboratorsPresenceProps {
  * Renders a list of avatars for the active collaborators, with a maximum of 3 visible avatars.
  * Shows a popover with all collaborators on hover.
  *
- * @param {Object} props          CollaboratorsPresence component props
- * @param {number} props.postId   ID of the post
- * @param {string} props.postType Type of the post
+ * @param props          CollaboratorsPresence component props
+ * @param props.postId   ID of the post
+ * @param props.postType Type of the post
  */
 export function CollaboratorsPresence( {
 	postId,
@@ -52,51 +59,51 @@ export function CollaboratorsPresence( {
 		return null;
 	}
 
-	const visibleCollaborators = otherActiveCollaborators.slice( 0, 3 );
-	const remainingCollaborators = otherActiveCollaborators.slice( 3 );
-	const remainingCollaboratorsText = remainingCollaborators
-		.map( ( { collaboratorInfo } ) => collaboratorInfo.name )
-		.join( ', ' );
-
-	return visibleCollaborators.length > 0 ? (
-		<div className="editor-collaborators-presence">
-			<Button
-				__next40pxDefaultSize
-				className="editor-collaborators-presence__button"
-				onClick={ () => setIsPopoverVisible( ! isPopoverVisible ) }
-				isPressed={ isPopoverVisible }
-				ref={ setPopoverAnchor }
-				aria-label={ sprintf(
-					// translators: %d: number of online collaborators.
-					__( 'Collaborators list, %d online' ),
-					otherActiveCollaborators.length
-				) }
-			>
-				{ visibleCollaborators.map( ( collaboratorState ) => (
-					<Avatar
-						key={ collaboratorState.clientId }
-						collaboratorInfo={ collaboratorState.collaboratorInfo }
-						showCollaboratorColorBorder={ false }
-						size="small"
+	return (
+		<>
+			<div className="editor-collaborators-presence">
+				<Button
+					__next40pxDefaultSize
+					className="editor-collaborators-presence__button"
+					onClick={ () => setIsPopoverVisible( ! isPopoverVisible ) }
+					isPressed={ isPopoverVisible }
+					ref={ setPopoverAnchor }
+					aria-label={ sprintf(
+						// translators: %d: number of online collaborators.
+						__( 'Collaborators list, %d online' ),
+						otherActiveCollaborators.length
+					) }
+				>
+					<AvatarGroup max={ 3 }>
+						{ otherActiveCollaborators.map(
+							( collaboratorState ) => (
+								<Avatar
+									key={ collaboratorState.clientId }
+									src={ getAvatarUrl(
+										collaboratorState.collaboratorInfo
+											.avatar_urls
+									) }
+									name={
+										collaboratorState.collaboratorInfo.name
+									}
+									borderColor={ getAvatarBorderColor(
+										collaboratorState.collaboratorInfo.id
+									) }
+									size="small"
+								/>
+							)
+						) }
+					</AvatarGroup>
+				</Button>
+				{ isPopoverVisible && (
+					<CollaboratorsList
+						activeCollaborators={ otherActiveCollaborators }
+						popoverAnchor={ popoverAnchor }
+						setIsPopoverVisible={ setIsPopoverVisible }
 					/>
-				) ) }
-
-				{ remainingCollaborators.length > 0 && (
-					<div
-						className="editor-collaborators-presence__remaining"
-						title={ remainingCollaboratorsText }
-					>
-						+{ remainingCollaborators.length }
-					</div>
 				) }
-			</Button>
-			{ isPopoverVisible && (
-				<CollaboratorsList
-					activeCollaborators={ otherActiveCollaborators }
-					popoverAnchor={ popoverAnchor }
-					setIsPopoverVisible={ setIsPopoverVisible }
-				/>
-			) }
-		</div>
-	) : null;
+			</div>
+			<CollaboratorsOverlay postId={ postId } postType={ postType } />
+		</>
+	);
 }